refactored web-scan connections
